outsource methods in utils and adjust styling
